Reorganize cruftfiles postboxes into security page
